Refactor other macros
