+++ packages/babel-plugin-emotion/src/styled-macro.js
@@ -1,31 +1,11 @@
 // @flow
 import { createMacro } from 'babel-plugin-macros'
-import { addDefault, addNamed } from '@babel/helper-module-imports'
-import { transformExpressionWithStyles, getStyledOptions } from './utils'
-
-function addImport(
-  state: any,
-  importPath: string,
-  imported: string,
-  nameHint?: string
-) {
-  let cacheKey = ['import', importPath, imported].join(':')
-  if (state[cacheKey] === undefined) {
-    let importIdentifier
-    if (imported === 'default') {
-      importIdentifier = addDefault(state.file.path, importPath, { nameHint })
-    } else {
-      importIdentifier = addNamed(state.file.path, imported, importPath, {
-        nameHint
-      })
-    }
-    state[cacheKey] = importIdentifier.name
-  }
-  return {
-    type: 'Identifier',
-    name: state[cacheKey]
-  }
-}
+import { addNamed } from '@babel/helper-module-imports'
+import {
+  transformExpressionWithStyles,
+  getStyledOptions,
+  addImport
+} from './utils'
 
 export let styledTransformer = ({
   state,
