+++ packages/babel-plugin-emotion/src/utils/index.js
@@ -9,3 +9,4 @@ export {
 } from './transform-expression-with-styles'
 export { getStyledOptions } from './get-styled-options'
 export { appendStringToExpressions, joinStringLiterals } from './strings'
+export { addImport } from './add-import'
